third_party/blink/web_tests/external/wpt/css/css-shapes/shape-functions/ellipse-function-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['shape-outside'] = "ellipse(20px 40px at center)" should set the property value
  assert_equals: serialization should be canonical expected "ellipse(20px 40px at center center)" but got "ellipse(20px 40px at 50% 50%)"
[FAIL] e.style['shape-outside'] = "ellipse(closest-corner 20%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "ellipse(farthest-corner 20%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "ellipse(4% 20% at top right)" should set the property value
  assert_equals: serialization should be canonical expected "ellipse(4% 20% at right top)" but got "ellipse(4% 20% at 100% 0%)"
[FAIL] e.style['shape-outside'] = "ellipse(10px closest-side at top right)" should set the property value
  assert_equals: serialization should be canonical expected "ellipse(10px closest-side at right top)" but got "ellipse(10px closest-side at 100% 0%)"
[FAIL] e.style['shape-outside'] = "ellipse(farthest-side 20px at center top)" should set the property value
  assert_equals: serialization should be canonical expected "ellipse(farthest-side 20px at center top)" but got "ellipse(farthest-side 20px at 50% 0%)"
[FAIL] e.style['shape-outside'] = "ellipse(farthest-side farthest-side at top right)" should set the property value
  assert_equals: serialization should be canonical expected "ellipse(farthest-side farthest-side at right top)" but got "ellipse(farthest-side farthest-side at 100% 0%)"
Harness: the test ran to completion.
